refactor: Move Oauth2 and useProjects to react-hooks package
