+++ sdks/react-hooks/package.json
@@ -20,7 +20,6 @@
   "dependencies": {
     "axios": "^0.27.2",
     "ts-node": "^10.9.1",
-    "@kiqr/react": "workspace:*",
     "@kiqr/management-api-sdk": "workspace:*",
     "react": "^18.2.0",
     "swr": "^1.3.0"
