+++ sdks/react-hooks/src/hooks/use-project.ts
@@ -1,6 +1,10 @@
+import { useContext } from 'react'
+import { KiqrContext } from '../kiqr-context'
 import { Configuration, Project, ProjectsApi } from '@kiqr/management-api-sdk'
 
-export const getProject = async (
+import useSWR from 'swr'
+
+const getProject = async (
   accessToken: string,
   projectId: string
 ): Promise<Project> => {
@@ -13,4 +17,14 @@ export const getProject = async (
       .then((response) => resolve(response.data))
       .catch((error) => reject(error.message))
   })
+}
+
+export const useProject = (projectId: string) => {
+  const { token } = useContext(KiqrContext)
+  const { data: project, error: projectsError } = useSWR(
+    [token?.access_token, projectId],
+    (token, projectId) => getProject(token as string, projectId as string)
+  )
+
+  return { project, projectsError, token }
 }
