+++ sdks/react-hooks/src/hooks/use-projects.ts
@@ -1,4 +1,8 @@
+import { useContext } from 'react'
+import { KiqrContext } from '../kiqr-context'
 import { Configuration, Project, ProjectsApi } from '@kiqr/management-api-sdk'
+
+import useSWR from 'swr'
 
 export const getProjects = async (accessToken: string): Promise<Project[]> => {
   const configuration = new Configuration({ accessToken: accessToken })
@@ -10,4 +14,14 @@ export const getProjects = async (accessToken: string): Promise<Project[]> => {
       .then((response) => resolve(response.data))
       .catch((error) => reject(error.message))
   })
+}
+
+export const useProjects = () => {
+  const { token } = useContext(KiqrContext)
+  const { data: projects, error: projectsError } = useSWR(
+    [token?.access_token, '/projects'],
+    getProjects
+  )
+
+  return { projects, projectsError, token }
 }
