+++ sdks/react-hooks/src/kiqr-context.ts
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Oauth2Token } from '.'
+import { Oauth2Token } from './oauth2-config'
 
 export interface KiqrContextConfig {
   token: Oauth2Token | null
