+++ sdks/react/package.json
@@ -65,6 +65,7 @@
   },
   "dependencies": {
     "@kiqr/management-api-sdk": "workspace:*",
+    "@kiqr/react-hooks": "workspace:*",
     "axios": "^0.27.2",
     "swr": "^1.3.0"
   },
