--- sdks/react/src/hooks/index.ts
@@ -1,6 +1,0 @@
-export { useEnvironments } from './use-environments'
-export { useProject } from './use-project'
-export { useProjects } from './use-projects'
-export { useSchema } from './use-schema'
-export { useSchemas } from './use-schemas'
-export { useSession } from './use-session'
