+++ sdks/react/src/index.ts
@@ -11,11 +11,3 @@ export { Padding } from './components/atoms/Padding'
 export { Placeholder } from './components/atoms/Placeholder'
 export { ProgressBar } from './components/atoms/ProgressBar'
 export { AvatarStack } from './components/molecules/AvatarStack'
-
-// Hooks
-export * from './hooks'
-
-// Logical components / Helpers
-export { KiqrContext, KiqrContextConfig } from './kiqr-context'
-export { KiqrProvider, KiqrProviderProps } from './kiqr-provider'
-export { Oauth2Config, Oauth2Token } from './oauth2-config'
