--- sdks/react/src/requests/index.ts
@@ -1,6 +1,0 @@
-export { getEnvironments } from './get-environments'
-export { getProject } from './get-project'
-export { getProjects } from './get-projects'
-export { getSchema } from './get-schema'
-export { getSchemas } from './get-schemas'
-export { getUser } from './get-user'
